==> code/07_beta_diversity_tilapia.Rmd <==
---
title: "Beta diversity"
author: "Yanxian Li<br><small>Department of Paraclinical Sciences<br>Faculty of Veterinary Medicine<br>Norwegian University of Life Sciences</small>"
date: "<small>`r Sys.Date()`</small>"
output:
  html_document:
    toc: yes
    toc_depth: 3
    number_sections: true
    toc_float: 
      collapsed: true
    code_folding: show
    theme: cerulean
    self_contained: true
  pdf_document: 
    latex_engine: xelatex
  word_document: default  
---

```{r style, echo = FALSE, message = FALSE, warning = FALSE}
require(knitr)
options(width = 120)
opts_chunk$set(message = FALSE, error = FALSE, warning = FALSE, cache = FALSE)
```

`r Hmisc::hidingTOC(buttonLabel = "Outline (hide)", tocSide = "right",  buttonSide = "right")`

# Getting ready

Load packages

```{r}
library(here) # A Simpler Way to Find Your Files, CRAN
library(tidyverse) # Easily Install and Load the 'Tidyverse', CRAN
library(patchwork) # The Composer of Plots, CRAN
library(flextable) # Functions for Tabular Reporting, CRAN
library(usedist) # Distance Matrix Utilities
library(vegan) # Community Ecology Package, CRAN
```

Load function.

```{r}
source(here("code/functions/plot_betadisper.R"))
```

Load data

```{r}
# metadata 
mtd <- read_tsv(here("data/metadata.tsv"), comment = "#q2") %>%
  rename(SampleID = "#SampleID") %>%
  filter(Species == "Nile tilapia") %>%
  mutate(
    Population = factor(Population, levels = c("Farmed", "Wild")),
    Environment = factor(Environment, levels = c("Pond", "Lake"))
  )

# distance matrix
dist <- readRDS(here("data/intermediate/qiime2R/tilapia/beta_diversity_distance.rds"))
dist_atch <- dist$aitchison$data # Aitchison distance 
dist_wuf <- dist$weighted_unifrac$data # weighted UniFrac distance 

# ordination 
ord <- readRDS(here("data/intermediate/qiime2R/tilapia/beta_diversity_ordination.rds")) 
ord_atch <- ord$aitchison$data # Aitchison distance based PCoA
ord_wuf <- ord$weighted_unifrac$data # weighted UniFrac distance based PCoA
```

# Ordination

Unweighted beta-diversity distance metrics such as unweighted UniFrac are subjective to technical noises, making comparison of results across studies difficult ([Wong et al., 2016](https://doi.org/10.1371/journal.pone.0161196); [Nearing et al., 2018](https://doi.org/10.7717/peerj.5364)). Here we use weighted distance metrics including robust Aitchison distance and weighted UniFrac distance to compare beta-diversity. Robust Aitchison distance is a compositional beta diversity metric rooted in a centered log-ratio (CLR) transformation and matrix completion [(Martino et al., 2019)](https://msystems.asm.org/content/4/1/e00016-19). It has some desirable properties: 1)scale invariant, which ensures equivalence between distances computed from absolute and relative abundance measurements, negating the need to perform rarefaction; 2)relative changes driven. Microbes that display large fold change across samples will be weighted more heavily, which makes the ordination robust to random fluctuations of high-abundant taxa; 3)sub-compositionally coherent, which guarantees that distances will never decrease if additional taxa are observed.

## PCoA: robust Aitchison distance

```{r}
pcoa_atch <- ord_atch$Vectors %>% 
  mutate(
    SampleID = as.character(SampleID),
    SampleID = ifelse(nchar(SampleID) < 2, paste0("0", SampleID), SampleID)
    ) %>% 
  inner_join(mtd, by = "SampleID") %>%
  ggplot(aes(x = PC1, y = PC2, color = Population, shape = Environment)) +
    geom_hline(yintercept = 0, color = "grey") +
    geom_vline(xintercept = 0, color = "grey") +
    # data ellipse
    stat_ellipse(linetype = "dashed") +
    geom_point(size = 2) +
    # keep the aspect ratio between y and x axis (y-axis/x-axis)
    coord_fixed(ord_atch$ProportionExplained[2]/ord_atch$ProportionExplained[1]) + 
    labs(
      x = paste0("PCo1: ", round(100 * ord_atch$ProportionExplained[1], 1), "%"),
      y = paste0("PCo2: ", round(100 * ord_atch$ProportionExplained[2], 1), "%")
      ) +
    scale_color_brewer(palette = "Set2") +
    theme_bw(base_size = 14)

pcoa_atch + labs(title = "PCoA of robust Aitchison distance")
```

## PCoA: weighted UniFrac distance

```{r}
pcoa_wuf <- ord_wuf$Vectors %>% 
  mutate(
    SampleID = as.character(SampleID),
    SampleID = ifelse(nchar(SampleID) < 2, paste0("0", SampleID), SampleID)
    ) %>% 
  inner_join(mtd, by = "SampleID") %>%
  ggplot(aes(x = PC1, y = PC2, color = Population, shape = Environment)) +
    geom_hline(yintercept = 0, color = "grey") +
    geom_vline(xintercept = 0, color = "grey") +
    stat_ellipse(linetype = "dashed") +
    geom_point(size = 2) +
    coord_fixed(ord_wuf$ProportionExplained[2]/ord_wuf$ProportionExplained[1]) + 
    labs(
      x = paste0("PCo1: ", round(100 * ord_wuf$ProportionExplained[1], 1), "%"),
      y = paste0("PCo2: ", round(100 * ord_wuf$ProportionExplained[2], 1), "%")
      ) +
    scale_color_brewer(palette = "Set2") +
    theme_bw(base_size = 14) 

pcoa_wuf + labs(title = "PCoA of weighted UniFrac distance")
```

## Figure 4

Assemble plots.

```{r, fig.width=10}
pcoa_atch + pcoa_wuf + 
  plot_layout(guides = "collect") +
  plot_annotation(tag_levels = "a")

# export plot
ggsave(here("result/figure/tilapia/Figure 4.tiff"), width = 10, units = "in", 
       dpi = 300, compression = "lzw")
```

# PERMANVOA

## Export distance matrix

We export the distance matrices and run PERMANOVA using the PERMANOVA+ add-on from the [PRIMER v7](https://www.primer-e.com/our-software/permanova-add-on/).

```{r, results='hide'}
# combine distance matrices into a list 
dist_list <- list(dist_atch, dist_wuf)
names(dist_list) <- c("Aitchison_distance", "Weighted_UniFrac_distance")

# export distance matrices
lapply(seq_along(dist_list), function(x) {
  dist_list[[x]] %>%
    as.matrix() %>%
    as.data.frame() %>%
    write.table( 
      here(
        paste0("data/intermediate/permanova/tilapia/", names(dist_list)[x], ".tsv")
        ), 
      sep = "\t", 
      col.names = NA, 
      row.names = TRUE)
    }
  )
```

## Import and format PERMANVOA results

```{r}
# import permanova results
permanova <- read_csv(
  here("data/intermediate/permanova/tilapia/permanova_results.csv")
  )

# format permanova results
permanova %>% 
  # the following 4 lines of code adjust raw p values
  group_nest(Distance, Comparison) %>%
  mutate(P_adj = map(data, ~p.adjust(.x$P, method = "holm"))) %>%
  unnest(cols = c(data, P_adj)) %>%
  ungroup() %>%
  mutate(
    Comparison = gsub(
      "Tilapia_farmed_pond - Tilapia_farmed_lake", 
      "farmed tilapia", 
      Comparison
      ),
    Comparison = gsub(
      "Tilapia_farmed_lake - Tilapia_wild_lake", 
      "tilapia in lake", 
      Comparison
      ),
    Comparison = factor(
      Comparison, levels = c("farmed tilapia", "tilapia in lake")
      ),
    Term = gsub("Environment", "Environment (Pond vs Lake)", Term),
    Term = gsub("Population", "Population (Farmed vs Wild)", Term),
    P = P_adj
    ) %>%
  select(-P_adj) %>%
  arrange(Distance, Comparison) %>%
  flextable() %>%
  set_caption(
    caption = "Table 1. PERMANOVA results based on robust Aitchison and weighted UniFrac distance."
    ) %>%
  autofit() %>%
  italic(j = 5:6, italic = TRUE, part = "head") %>% # italicize column names
  merge_v(j = c("Distance", "Comparison")) %>% # merge cells by phylum
  theme_booktabs() %>% # apply table style
  align(align = "left", part = "all") %>% # align text to the left
  hline(i = c(3, 6, 9), part = "body") # draw horizontal lines 
```

# PERMDISP

PERMANOVA tests differences in both location and dispersion effects. Although PERMANOVA is robust to heterogeneity in the multivariate dispersion, it's a good practice to look at the homogeneity of multivariate dispersion following a significant PERMANVOA result. The homogeneity of multivariate dispersion can be assessed visually (PCoA plot/boxplot) or by a permutation test called PERMDISP, which is implemented in R by the `betadisper()` function from the *vegan* package. When within-group sample sizes are small (n < 5), formal testing with PERMDISP is not very meaningful. Visual assessment is a better option.  

## Robust Aitchison distance

### Farmed tilapia: pond vs lake

Subset distance matrix.

```{r}
mtd_farmed <- filter(mtd, Population == "Farmed") %>%
  mutate(Site = paste0(Environment, "_", Site))

dist_atch_farmed <- dist_subset(dist_atch, mtd_farmed$SampleID) 
```

#### Site-level

First, we check the equality of dispersion for samples (replicates) across different Site (centroids).

Compute beta-dispersion.

```{r}
(
disp_atch_farmed_site <- betadisper(
  d = dist_atch_farmed, 
  group = mtd_farmed$Site, 
  type = "median", 
  bias.adjust = TRUE
  )
)
```

Visual inspection.

```{r, fig.width=6.5}
plot_betadisper(
  disp_atch_farmed_site, 
  xlab = "Site",
  title = "Multivariate dispersion: farmed tilapia (pond vs. lake)") +
  labs(subtitle = "robust Aitchison distance, site-level") +
  coord_flip()
```

#### Environment-level

Next, we consider differences in the dispersion of Site centroids around the Environment centroids.

Compute distance among the Site centroids.

```{r}
dist_atch_farmed_site <- dist_multi_centroids(dist_atch_farmed, mtd_farmed$Site)
```

Make a new metadata matching Site and Environment.

```{r}
mtd_farmed_site <- mtd_farmed %>%
  select(Environment, Site) %>%
  distinct()%>%
  arrange(Site) 
```

Compute beta-dispersion.

```{r}
(
disp_atch_farmed_env <- betadisper(
  d = dist_atch_farmed_site, 
  group = mtd_farmed_site$Environment, 
  type = "median", 
  bias.adjust = TRUE
  )
)
```

Visual inspection.

```{r, fig.width=6.5}
plot_betadisper(
  disp_atch_farmed_env, 
  xlab = "Environment",
  title = "Multivariate dispersion: farmed tilapia (pond vs. lake)") +
  labs(subtitle = "robust Aitchison distance, environment-level") 
```

### Tilapia in lake: farmed vs wild

Subset distance matrix.

```{r}
mtd_lake <- filter(mtd, Environment == "Lake") %>%
  mutate(Site = paste0(Population, "_", Site))

dist_atch_lake <- dist_subset(dist_atch, mtd_lake$SampleID) 
```

#### Site-level

Compute beta-dispersion.

```{r}
(
disp_atch_lake_site <- betadisper(
  d = dist_atch_lake, 
  group = mtd_lake$Site, 
  type = "median", 
  bias.adjust = TRUE
  )
)
```

Visual inspection.

```{r, fig.width=6.5}
plot_betadisper(
  disp_atch_lake_site, 
  xlab = "Site",
  title = "Multivariate dispersion: tilapia in lake (farmed vs. wild)") +
  labs(subtitle = "robust Aitchison distance, site-level") +
  coord_flip()
```

#### Population-level

Compute distance among the Site centroids.

```{r}
dist_atch_lake_site <- dist_multi_centroids(dist_atch_lake, mtd_lake$Site)
```

Make a new metadata matching Site and Population.

```{r}
mtd_lake_site <- mtd_lake %>%
  select(Population, Site) %>%
  distinct()%>%
  arrange(Site) 
```

Compute beta-dispersion.

```{r}
(
disp_atch_lake_pop <- betadisper(
  d = dist_atch_lake_site, 
  group = mtd_lake_site$Population, 
  type = "median", 
  bias.adjust = TRUE
  )
)
```

Visual inspection.

```{r, fig.width=6.5}
plot_betadisper(
  disp_atch_lake_pop, 
  xlab = "Population",
  title = "Multivariate dispersion: tilapia in lake (farmed vs. wild)") +
  labs(subtitle = "robust Aitchison distance, population-level") 
```

## Weighted UniFrac distance

### Farmed tilapia: pond vs lake

Subset distance matrix.

```{r}
dist_wuf_farmed <- dist_subset(dist_wuf, mtd_farmed$SampleID) 
```

#### Site-level

Compute beta-dispersion.

```{r}
(
disp_wuf_farmed_site <- betadisper(
  d = dist_wuf_farmed, 
  group = mtd_farmed$Site, 
  type = "median", 
  bias.adjust = TRUE
  )
)
```

Visual inspection.

```{r, fig.width=6.5}
plot_betadisper(
  disp_wuf_farmed_site, 
  xlab = "Site",
  title = "Multivariate dispersion: farmed tilapia (pond vs. lake)") +
  labs(subtitle = "Weighted UniFrac distance, site-level") +
  coord_flip()
```

#### Environment-level

Compute distance among the Site centroids.

```{r}
dist_wuf_farmed_site <- dist_multi_centroids(dist_wuf_farmed, mtd_farmed$Site)
```

Compute beta-dispersion.

```{r}
(
disp_wuf_farmed_env <- betadisper(
  d = dist_wuf_farmed_site, 
  group = mtd_farmed_site$Environment, 
  type = "median", 
  bias.adjust = TRUE
  )
)
```

Visual inspection.

```{r, fig.width=6.5}
plot_betadisper(
  disp_wuf_farmed_env, 
  xlab = "Environment",
  title = "Multivariate dispersion: farmed tilapia (pond vs. lake)") +
  labs(subtitle = "Weighted UniFrac distance, environment-level")
```

### Tilapia in lake: farmed vs wild

Subset distance matrix.

```{r}
dist_wuf_lake <- dist_subset(dist_wuf, mtd_lake$SampleID) 
```

As PERMANOVA suggests little evidence of Site effect, here we use individual samples for assessing homogeneity of multivariate dispersion at population-level.

#### Population-level

Compute beta-dispersion.

```{r}
(
disp_wuf_lake_pop <- betadisper(
  d = dist_wuf_lake, 
  group = mtd_lake$Population, 
  type = "median", 
  bias.adjust = TRUE
  )
)
```

Formal permutation test with `permutest`.

```{r}
permutest(disp_wuf_lake_pop, permutations = 999)
```

Visual inspection.

```{r, fig.width=6.5}
plot_betadisper(
  disp_wuf_lake_pop, 
  xlab = "Population",
  title = "Multivariate dispersion: tilapia in lake (farmed vs. wild)") +
  labs(subtitle = "Weighted UniFrac distance, population-level") 
```

# Session information

```{r session-info}
sessionInfo()
```
